+++ word-game-be/util/Ably.go
@@ -6,6 +6,6 @@ import (
 	"github.com/ably/ably-go/ably"
 )
 
-func PublishLobbyMessage(client *ably.Realtime, lobby *uint32, name string, message interface{}) error {
+func PublishLobbyMessage(client *ably.Realtime, lobby *int64, name string, message interface{}) error {
 	return client.Channels.Get(fmt.Sprintf("lobby-%d", *lobby)).Publish(context.Background(), name, message)
 }
